api/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.4
PyJWT==2.9.0
setuptools==80.9.0
sqlparse==0.5.3
